requirements.txt
inky[rpi,example-depends]==1.5.0
viam-sdk==0.13.3
pillow==10.2.0
